feat: rename SidebarContextProvider file
